requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
coloredlogs==15.0.1
future==0.18.2
graphviz==0.20
grpcio==1.45.0
humanfriendly==10.0
idna==3.3
networkx==2.6.3
protobuf==3.20.1
py-solc-x==1.1.1
pyevmasm==0.2.3
pygraphviz==1.7
PyYAML==6.0
requests==2.27.1
semantic-version==2.9.0
six==1.16.0
urllib3==1.26.9
z3-solver==4.8.17.0
